+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.1.0
+
+* Added podcast image
+* Allow setting URL 
+* Expose client library
+ 
 ## v1.0.0
 
 * Introducing podcast feed for narrated articles (/podcast)